requirements.txt
pyserial>=3.5
pygame>=2.0.0



